+++ core/serializers.py
@@ -8,6 +8,10 @@ from rest_framework import serializers
 
 class ProductLookupSerializer(serializers.Serializer):
     q = serializers.CharField(required=False)
+    tx_id = serializers.CharField(required=False)
+    interraction_id = serializers.CharField(required=False)
+    value_id = serializers.CharField(required=False)
+    value_string = serializers.CharField(required=False)
 
 
 class CompanySerializer(serializers.Serializer):