+++ react-components/src/bundle.js
@@ -10,6 +10,8 @@ import ProductLookup from '@src/views/ProductLookup/Container'
 import LoginModal from '@src/views/LoginModal/Modal'
 import MarketSelectNavbar from '@src/views/MarketSelectNavbar/Container'
 import Tour from '@src/views/Tour/Container'
+import ProductFinder from '@src/components/ProductFinder/ProductFinder'
+import CountryFinder from '@src/components/ProductFinder/CountryFinder'
 import { createRouteToMarket, createSpendingAndResources } from '@src/views/sections/MarketingApproach'
 import { aboutYourBusinessForm } from '@src/views/sections/AboutYourBusiness'
 import { adaptToTargetMarketForm } from '@src/views/sections/AdaptationForYourTargetMarket'
@@ -38,6 +40,8 @@ export default {
   IndustriesModal,
   LearnIntroduction,
   ProductLookup,
+  ProductFinder,
+  CountryFinder,
   LoginModal,
   MarketSelectNavbar,
   ProductsModal,