+++ CHANGELOG.md
@@ -50,7 +50,9 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-208 - Product search and country search modal
 - GP2-259 - Lesson - pros and cons (BE)
+- GP2-208 - Product search modal
 - GP2-259 - Lesson - pros and cons (FE)
 - GP2-336 - Lesson progress on dashboard
 - GP2-386 - Lesson - hero image (FE) - Resize